chatproxy/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
python-jose[cryptography]>=3.3
pydantic-settings>=2.2
cachetools>=5.3
redis[asyncio]>=5.0
pytest>=8.0
pytest-asyncio>=0.23
